editing varibale types for C90

+++ 0x02-search_algorithms/0-linear.c
@@ -15,11 +15,11 @@ int linear_search(int *array, size_t size, int value)
 
 	if (array != NULL)
 	{
-		size_t i;
+		unsigned int i;
 
 		for (i = 0; i < size; i++)
 		{
-			printf("Value checked array[%zu] = [%i]\n", i, array[i]);
+			printf("Value checked array[%d] = [%d]\n", i, array[i]);
 			if (array[i] == value)
 			{
 				return (i);
